Extract Store::{next_scheduled, promote_scheduled} etc.
Following on from fd884206fb4b05dd182a90a903d5038d152b28fc this time
with job-scheduling related methods, further breaking down `Store`.
